Changed task component name to column card
