+++ components/ColumnCard/styles.module.scss
@@ -1,4 +1,4 @@
-.task {
+.columncard {
   margin: 10px 0;
   position: relative;
   width: 100%;
